Fixed issue with integer overflow on 64bit systems
See http://mattgemmell.com/2008/05/20/mgtemplateengine-templates-with-cocoa/comment-page-1#comment-42393

+++ ICUTemplateMatcher.m
@@ -151,7 +151,7 @@
 	NSString *argsPattern = @"\"(.*?)(?<!\\\\)\"|'(.*?)(?<!\\\\)'|(\\S+)";
 	NSMutableArray *args = [NSMutableArray array];
 	
-	int location = 0;
+	NSUInteger location = 0;
 	while (location != NSNotFound) {
 		NSRange searchRange  = NSMakeRange(location, [argString length] - location);
 		NSRange entireRange = [argString rangeOfRegex:argsPattern options:RKLNoOptions
